Handle transparencies with dynamic geometry events
When a dynamic mesh event happens, molecules can end up in new meshes.
Originally, we would always move molecules back to whatever was their
closest enclosing mesh before the event happened. It makes more sense to
let molecules pass through meshes that they are transparent to.

This commit adds partial support for this. If the molecule moves to an
"overlapping" location, then the molecule will move through meshes it is
transparent to. By overlapping, we mean that a molecule moved into or
out of a single set of nested meshes. For example, imagine mesh A is
inside mesh B which inside C (A->B->C->null). If the closest enclosing
mesh was A and is now C or vice versa, that is considered an overlapping
movement.