Handle transparencies w/ "non-overlapping" meshes.
When a dynamic mesh event happens, molecules can end up in new meshes.
Molecules should be able to pass through meshes that they are
transparent to during such an event.

A recent commit added partial support for this, namely if meshes
"overlapped". See that commit for an explanation of what is meant by
overlapping. With this commit, we can handle molecule movements between
non-overlapping meshes like A->B->C->null to D->E->F->null.