Removed/commented dead code.
Code that was reported unreachable by clang was removed or commented
out. There were many sections with return/break/free statements which
couldn't be reached because of calls to mcell_error or some similar
function, which killed MCell before it could ever get to it. In a few
cases, some of the cleanup code was moved ahead of the fatal errors, so
that it was actually being called before the program quit. Simply
removing the code is likely not the best long term solution, as we might
want to be able to recover from some of these failures if they are used
in API calls. However, for the time being, there doesn't seem to be any
reason to leave in this dead code, since it's pointless and misleading
at best.