Updated wall connectivity policy to take into account regions.
Previously, for geometry objects with edges connecting more than
two walls we picked those two walls with the most divergent normals.
However, occasionally it is advantageous to give preference to walls
which share a common region even if they are not the ones with the
most divergent normals. This allows, e.g., to reliably count on
or release inside (closed) regions which are part of an overall
non-manifold object (e.g. a cube with triangles sticking out at
various angles).