Add error handling if meshes are added or removed.
With dynamic geometry, the assumption is that even though the geometry
changes (i.e deformations, adding/removing faces, etc), no new objects
are added or removed between events. For example if a user instantiated
one single mesh called Cube, then that mesh and only that mesh must
appear in future dynamic geometry events. However, there was no test in
place to actually check if this was true, and so it would result in
strange and unexpected errors if a user accidentally instantiated new
objects or forgot to re-instantiate previous objects. A test has been
added to verify that this is true. We might want to loosen this
restriction in the future, so the implemenation is simple, isolated and
doesn't touch much of the other dynamic geometry code.

In investigating this, another small bug was fixed. If the order of mesh
re-instantiations was switched around, this would trigger an error,
because the root instance wasn't being properly cleaned up.