Added additional surface clamp and surface region code.
This commit also adds experimental syntax to deal with both
the orientation of surface molecules and the sidedness of
special surface reactions. E.g., for clamps of surface molecules
we need to specify both the orientation of clamped molecules and
the side with respect to the region edge on which to clamp
molecules. Similarly, for absorptive/reflective/transparent
surface regions we need to specify both the side which has this
property as well as the orientation of surface molecules it applies
to. The current version of the syntax is

CLAMP_CONCENTRATION A, = 1e-5 {1}

or

REFLECTIVE = A' {-1}

Here, analogous to the volume case, the orientation mark on the molecule
itself specifies the side of the region edge which has this property.
The -1/0/1 in curly braces after the statement specifies the molecule
orientation. Using this syntax leaves our previous MDL for volume clamps
and absortive/reflective/transparent regions intact. Of course, we can
refine this syntax in the future once we have more experience with it.