Added Travis CI yaml file for build status.
This commit adds a yaml file needed by Travis CI (travis-ci.org), which is a
tool for reporting the build status of github projects. A git push triggers the
project to build and reports the status by email. The current status is also
typically added to the README.MD/RST file so it can quickly be checked at a
glance by users/developers.