Handle added or removed meshes with dyn. geoms.
Previously, if a mesh was added or removed during a dynagmic geometry
event, MCell just spit out an error message and quit. Now we allow
objects to be added or removed with the assumption that they will be
ignored when trying to place molecules according to the normal dynamic
geometry placement rules.

There are a few nuances to this that are probably easiest to explain
with an example. Imagine you have some volume molecules inside mesh "A".
During a dynamic geometry event, we remove "A" and add in "B" exactly
where "A" was. Normally, we would try to place molecules back in their
original mesh ("A"), but we will ignore it in this case since it has
been removed. Also, since "B" is now where "A" was, all of the molecules
are inside "B". In the normal case, we would also try to move molecules
outside of a mesh that they had moved into during a DG event, but we
ignore this as well since it's a brand new mesh. We only ignore meshes
that have just been added or removed in the most recent DG event. As
such, if another DG event happens that includes "B" again, we will pay
attention to it this time, since it was used in the previous DG event.