For surface molecules added ability to reflect/absorb from
a surface region boundary selectively when either entering
or leaving the region (or both). Also changed the syntax
to describe surface region properties to

CLAMP_CONCENTRATION A, = 1e-5 |1|

or

REFLECTIVE = A' |-1|

Previously, I used braces instead of the ||. However, since the
optional tick mark on the molecule name can also be represented
by braces this syntax was not unique.