Fixed bug in concentration clamp.
Previously, each surface class with a CONCENTRATION_CLAMP
statement applied to an object would have a single scaling factor
that took into account the sidedness of the clamp. Unfortunately,
this approach fails if multiple clamps are specified per surface
class which differ in their sidedness (e.g., one might be applied
to both sides of the surface, the other single sided). In this
case only one of the clamps was guaranteed to be scaled correctly,
the others could be of by a factor of 2 or 0.5. This was fixed
by removing the sidedness from the global scaling and rather apply
it once we compute the number of released molecules per species.