Changed how volume molecule placment works for DG.
Before, when inserting volume molecules into the world during a dynamic
geometry event, we would place them slightly offset from the center of
the closest appropriate wall. This can result in heavy clustering of
molecules. This commit makes it work similar to a concentration clamp;
in other words, we place them slightly offset from a random position on
the wall. It's not clear if this is the best solution, but it doesn't
have the clustering problem of the original approach.