Fixed bug in surface region code.
Walls on surfaces can have an arbitrary number of surface
classes assigned to them. In order to check that a wall
has a reactive region boundary (i.e. one that absorbs or
reflects molecules) we need to determine all regions with
a reflective/absorptive surface class that the wall is
part of and then check if the wall is a region boundary
for any of these regions. Previously, the code would
only consider the first reflective/absorptive surface
class. Thus, in the case of multiple overlapping reactive
surface classes it would not properly detect existing
reactive region boundaries. This could lead to leaks
across boundaries or missed absorption events. This
commit fixes this issue.