Dynamic geoms: surf molecs respect transparencies
Previously, MCell would check if all surface regions matched when trying
to place a molecule after a dynamic geometry event. For example, assume
molecule "sm" was previously on regions A and B. After moving meshes,
MCell would try to put molecule "sm" back on a wall that contained A and
B. The new wall it was placed on couldn't be missing any regions (A or
B) and it couldn't contain any extra regions (C, D, etc).

Now, if one of the regions is transparent during a dynamic geometry
event, MCell will allow the molecule to move into or out of those
regions. However, it doesn't distinguish between surface molecule
orientations properly yet.