Bugfix: birthdays wrong when checkpointing
If a user changed the timestep between simulations via checkpointing, the
birthdays and cumulative lifetimes of molecules would be handled incorrectly,
since they were both treated as timesteps/iterations. Note: a recent commit
changed how this worked slightly by always saving and reading the correct time.
This was better, but it didn't fix the issue.

It's easiest to see why the original approach was a problem with an example.
Imagine a molecule is born at 50 iterations into a 100 iteration simulation run
with a timestep of 1e-6s. This means the molecule was born at 50 microseconds.
A checkpoint file is created with the birthday set to 50. When a new simulation
starts, we read in that birthday as 50. However, our simulation now runs with a
timestep of 1e-7s, and so our birthday of 50 now means 5 microseconds.

This could result in incorrect reporting of mean lifetimes. The new method
avoids the problems of the previous method by treating both birthday and
cumulative lifetimes as seconds. cum_lifetimes was also renamed
cum_lifetimes_seconds for clarity.

Lastly, some new functions were created to help deal with conversions from
iterations/timesteps to seconds and vice versa.