Add command line flag for "exact checkpoints"
If this flag/switch/option is set, it ensures the results from a checkpointed
simulation are the same as those generated from a non-checkpointed simulation.
For example, assume you have two identical simulations (aside from the
checkpoint iterations value). The one is checkpointed after 100 iterations.
After this, the checkpoint file is loaded and the simulation runs for another
100 iterations (200 total). Then imagine you run a separate but essentially
identical simulation for 200 iterations without checkpointing. With this flag
set, the results will be identical for the checkpointed simulation (100+100
iters) and the non-checkpointed simulation (200 iters). NOTE: This option
should ONLY be set if the rate constants for a unimolecular reaction do NOT
change from one checkpointed simulation to the next.

If this flag is NOT set, the lifetimes for unimolecular reactions will be
recomputed after loading the checkpoint file, meaning the results for the
example above will not be identical. By default the flag is not set, since it
seems more important that the simulation results accurately reflect the rate
constants chosen rather then exactly matching the results of a non-checkpointed
version of the simulation.