Bugfix: rescheduling after checkpoint.
There was a problem in the previous commit when not using the
"exact_checkpoint" flag. The rescheduling wasn't working properly for
non-diffusing molecules.

The original logic was flawed on two accounts. There was no need to check if
the lifetime was greater than zero before setting the reschedule flag as
suggested in the deleted comment. Also, the the scheduling "time" (t) should
have been set to be the same as the current iterations, so it would get
rescheduled immediately. Otherwise, you'll get a much slower decay with
non-diffusing molecules.

NOTE: The "exact_checkpoint" flag will be removed in the next commit, since
there is actually no easy way to ensure that it is exact under all
circumstances. This bugfix is simply here in case we want to revisit using this
flag in the future.