Remove exact flag. Always reschedule after chkpt.
This commit removes the "exact_checkpoint" flag, since we cannot always ensure
that results from a checkpointed simulation will be the same as an otherwise
identical simulation run without checkpoints.

For instance, if you have diffusing surface molecules undergoing a simple
unimolecular reaction, the results will not be identical between the
checkpointed and non-checkpointed versions and there are likely other cases
too. Also, this method is completely broken if rates changes between
checkpoints, although we already knew about that limitation.

For now, we will simply always reschedule molecules when resuming a checkpoint.
The results will not be identical to those generated without checkpointing, but
the mean results should still be the same.