Usage message for logfreq default value was wrong.
The default value for the log frequency listed in the "usage message"
was listed as having a default value of 100, but the default value is
actually a function of how many total iterations are being run. Since
it's hard to explain all that in such a short space, the default message
was removed. It would be better to explain this in the QRG or a
tutorial.