Changed floating point comparison.
Simply looking at the code, it wasn't clear at all what the purpose of
checking if t2 equaled zero at this particular location in run_timestep.
Normally, a t2 of 0 means that a molecule is ready for a reaction.
However, as far as I can tell, the only reason to check it at this
section in the code was to skip over inert molecules.

So the question is why we should change it from using the
distinguishable function to merely checking if t2==0 exactly. It turns
out using the distinguishable function was especially prone to
triggering this test if you had a unimolecular reaction with a really
large rate constant (e.g. 1e30).

Also, it's not clear why we don't just set t to be equal to the total
number iterations instead of MAX_UNI_TIMESKIP (100000).

It's probably apparent at this point that this commit is somewhat of a
bandaid and a more comprehensive solution should be investigated. It
might be worth resurrecting the ACT_INERT flag for this purpose.