Bugfix: don't set previous_wall for static molecs.
The main reason for tracking the previous_wall is to prevent the
molecule from reacting with the wall it was produced from in the same
time step, which is particularly important for concentration clamps.
However, this shouldn't be an issue for non-diffusing molecules, since
they can't actually diffuse into the wall and react. As such, we simply
set the previous_wall to NULL if the diffusion constant is 0.

This is especially important to do when checkpointing, since it is
currently a fatal error to have a non-NULL previous_wall while writing
the checkpoint. As a side note, the reason this is an error is because
we don't currently save any information about the geometry in the
checkpoint file and simply setting it to NULL could bias the results.