Bugfix in COUNT code using ALL_ENCLOSED.
A user reported negative surface molecule counts when using the
ALL_ENCLOSED keyword. This behavior could be triggered reliably
by decreasing the partition size and thus the size of subvolumes.
The issue was due to a bug in collide_sv_time which would report
incorrect times for the next collision with subvolume boundaries. This
in turn would lead MCell to count certain wall crossings multiple
times screwing up the logic for deciding if a molecule is enclosed
in a volume or not.