Bugfix: Removing all objects.
If all the polygon objects are removed, then the world diagonal length
is set to zero. This length is used to calculate a random vector which
is used in raytracing. If this is a zero vector, then the collsion type
flag (smash->what) will be set to an invalid value. This ultimately
blows up when we try to traverse the subvolume.

The fix in this commit just checks if the world diagonal length is zero,
and changes it to some nonzero value.