Refactor: renamed variable (m->vm)
Volume molecules were renamed from m to vm (or vmp) in diffuse.c. This
naming format is already being used throughout other portions of the
code and is consistent with how abstract and surface molecules are named
(am and sm respectively). Also, it prevents false positives when
searching through the code base, since m is often used for 3D vectors.