Bugfix: events scheduled incorrectly
Due to a floating point precision problem, multiple events were being scheduled
for the same timestep. Maybe this should be allowed, but the way it was being
done was definitely wrong. If the event time was near an integer boundary like
6.99999, the event would get scheduled for the 6th (instead of the 7th) time
step due to an integer cast in sched_util.c.

Originally, I assumed this was causing a problem where molecules would leak out
of the mesh occasionally. It turns out that this was not the cause or at least
not the only one.