Replaced random vector code (rand -> rng_dbl).
The random vector being generated was using rand instead of rng_dbl.
Also, the random vector wasn't even normalized, so when it was scaled by
the world length, it wouldn't always span the entire world.

This code was used to find out what meshes a molecule was in via a
point-in-polygon test. As such, it could potentially misreport the
result if the vector was too small to cross the world.

For instance, a molecule which was actually in a mesh could be reported
as being inside nothing or a molecule outside of a mesh could be
reported as being inside of it like this:

(zero crossings)
    --------------
    |            |
    |            |
    |  *----->   |
    |            |
    |            |
    --------------

(one crossing)
    --------------
    |            |
    |            |
 *-------->      |
    |            |
    |            |
    --------------

instead of the correct solution:

(one crossing)
    --------------
    |            |
    |            |
    |  *------------->
    |            |
    |            |
    --------------

(two crossings)
    --------------
    |            |
    |            |
 *------------------->
    |            |
    |            |
    --------------

Note: It's not even clear why we need to use a random vector for this.
It would probably be just as effective (and more efficient) to use a
vector along one of the cardinal axes. I will try replacing this in a
future commit, but I want to save this fix in the commit history.