Bugfix: counting hits with dynamic geometries.
Counting hits (i.e. FRONT_HITS, BACK_HITS, ALL_HITS) was not working
correctly when using dynamic geometries, because a flag was being
incorrectly set. The current solution is probably still not ideal, but
it's an improvement over the previous hack. Also, two superfluous
functions were removed in the process.