Change mesh addition/removal to warning.
When meshes are added or removed (instead of merely changed) during a
dynamic geometry event, it is now treated as a proper warning instead of
just logging it to standard out. Like other warnings, it can be set to
IGNORED, WARNING, or ERROR. The default setting is WARNING, although we
might want to consider changing it to ERROR. The option itself is called
ADD_REMOVE_MESH. It might also be a good idea to have a final warning
after the simulation is finished in the same way we do for molecules
with small lifetimes.