Mark invalid/undefined counts appropriately.
If an object disappears during a dynamic geometry event, then a count
statement for that object is techincally undefined. In the past, the
counts would be reported as 0. However, it arguably makes more sense to
use some symbol to indicate that the count is in fact undefined. "X" was
arbitrarily chosen, but it might be changed to "NA", "-", or something
else. Alternatively, we could just let the user define whatever
arbitrary string they wanted to use.