This is the first commit to the smp_new branch.
The purpose of this branch is to gradually merge the "old" pthreads
based SMP code into your current master branch. Please note that
this is a huge patch. Even a git diff of the most recent smp code
from 2010 against the 2010 master is in excess of 10,000 lines.
It is probably at least twice that against the current master.
Thus, there is no way to merge the code automagically in a gracefully
manner and I have started to merge some bits by hand and re-implement
others on top of the current code base. This will probably be
a lengthy process and until it is done this branch will most likely
not even compile and if so certainly there should be no expectation
of it working correctly.