Initial parser for scanning dynamic geometry.
In order to be able to set up count statements for objects that don't
currently exist but will be added in a later dynamic geometry event, we
need to parse the geometry definitions and instantiations to make sure
the counts will be legal at some point in the future. The normal parser
wouldn't work for this purpose, so a custom parser and lexer have been
written to handle this task.

The parser in this initial commit doesn't do much yet and hasn't
actually been incorporated into the rest of the code base. It can be
manually compiled and run interactively like this:

lex meshlex.l
yacc -d meshparse.y
clang meshparse.c lex.yy.c y.tab.c -o meshparse
./meshparse

The only thing that it will currently parse is partition definitions,
like this:

PARTITION_X=[[0.1 TO 1.0 STEP 0.05]]