Added "instantiation" to dynamic geometry parser.
The dynamic geometry parser can now handle object instantiations. A
separate symbol table is created to store all the names. It's not
entirely clear yet how this should be incorporated into the main code
base yet. Nevertheless, something like this is needed when we are adding
count statements for objects that don't exist yet without polluting
vol->obj_sym_table.

There are still a number of limitations with the current dynamic
geometry parser, but it mostly works for simple cases. One thing that
needs to be fixed is verifying that instantiated objects have a
corresponding mesh object. Also, the yacc file needs organized,
comments need to be added to meshparse.c, and redundancies should be
removed.