Incorporate dyn. geom. parser into rest of code.
This is a first step at incorporating the dynamic geometry parser into
the rest of the code base. Currently, this is a bit of hackish mess,
partially attributed to not fully understanding the inner workings of
flex, bison, and Automake. More specifically, I couldn't find a good way
to have two separate parsers work with the Automake, so currently one
needs to generate lex.yy.c and y.tab.c which have been added to the
mcell_SOURCES of Makefile.am. Also, the main in y.tab.c is manually
commented out. This whole process should be automated or, better yet, an
alternative should be found.