Remove null pointer dereference in molecule species test.
After a molecule was destroyed in a bimolecular reaction we used
its species info for property testing. Instead of accessing the
molecule's properties directly (which is now defunct thus leading
to a null pointer dereference) we have to use a previously stored copy
instead.

NOTE: This part of the code is poorly designed and basically invites
these types of errors to happen.