Add significant improvements/bugfixes to SMP scheduling code.
This commit fixes a number of major bugs in the previous SMP
implentation which led to subvolumes not being scheduled properly.
As a result, the number of products created for a simple
reaction of type A + B -> C was significantly too low (factor
of 3 or so). With the current commit the number of reactants
is closer to the expected value but still incorrect
so additional problems in the SMP code remain to be identified
and fixed.

Also, previously the code contained several pthread_cond_wait
calls which were not properly protected with while loops
checking for their respective predicates. This was fixed
as well.