Count in/on regions that don't exist at beginning.
Previously, we only allowed counting in/on objects that didn't exist at
the beginning of the simulation, but now we can also do it in/on
regions. These objects/regions can be added later via a dynamic geometry
event.

This current solution is most likely overly complicated and will be
simplified in subsequent commits if at all possible.