Fix additional bugs in smp implementation.
One major issue which was fixed is that previously molecules
that would migrate to a new memory partition were not added
to the per species list of molecules in the relevant subvolumes
and thus would not participate in any reactions.

Also, migrating molecules were not properly marked as defunct
in the previous subvolume and could thus end up being
erroneously rescheduled.

This commit significantly improves MCell's ability to track
a simple reaction of type A + B -> C. However, the reaction
rate is still too slow compared to the ODE results (or
serial MCell for that matter). Thus, it looks like there
are more remaining issues that need to be tracked down
and fixed.