Bugfix: release patterns with DG
MCell could get stuck in an infinite loop when using release patterns
with dynamic geometries (if there wasn't a count statement for the
object in question). The problem was triggered by the fact that the
COUNT_CONTENTS flag was only being set in
mdl_set_release_site_geometry_object, which never gets called during a
DG event but this flag is tested for in find_enclosing_regions (called
from place_waypoints). The current fix is a bit of a hack, since ideally
we would only set this flag when it is needed, but we lose that
information during a DG event.