Treat initial geom the same as later DG events.
For the sake of consistency, the initial geometry and instantiation
block will now be treated the same way that geometry and instantiation
blocks are treated for every DG event. In other words, you would do
this:

--------------------------

  0 geom1.mdl
  5e-6 geom2.mdl
  10e-6 geom3.mdl

--------------------------

  DYNAMIC_GEOMETRY = "./geom.txt"

  INSTANTIATE Scene OBJECT
  {
    Release_Site RELEASE_SITE
    {
      SHAPE = DGScene.Cube
      MOLECULE = sm'
      NUMBER_TO_RELEASE = 100
    }
  }

instead of this:

--------------------------

  5e-6 geom2.mdl
  10e-6 geom3.mdl

--------------------------

  DYNAMIC_GEOMETRY = "./geom.txt"

  INCLUDE_FILE = "./geom1.mdl"

  INSTANTIATE Scene OBJECT
  {
    Cube OBJECT Cube {}
    Release_Site RELEASE_SITE
    {
      SHAPE = Scene.Cube
      MOLECULE = sm'
      NUMBER_TO_RELEASE = 100
    }
  }

Notice that there can still be an instantiation block in the main MDL for the
purpose of release sites.