Enable both forms of periodic boundary conditions.
This adds the form of periodic boundary conditions (traditional) from
the periodic_BC branch to this branch (periodic_BC_ng). In the
traditional form, volume molecules literally wrap around the box when
they hit the edge. In the non-traditional (periodic geometry) form, the
molecules reflect off the box thereby effectively entering what is a
virtual mirror image of the box. For now, there is no way for the user
to change between these options, and it is merely set with a bool in
reflect_or_periodic_bc. It is currently set to use the non-traditional
form. This will be made into a user selectable option in future commits.

Note: the terminology used in this commit (traditional, non-traditional,
periodic geometry) is completely subject to change if we think of better
ways to describe this.