Add custom PERIODIC_BOX syntax.
Instead of just adding extra keywords to a regular box object, this
creates a brand new keyword called PERIODIC_BOX. It's pretty similar to
the original method. In fact, behind the scenes, it's basically just
creating a box object object and instantiating it. The code could likely
be simplified, but for now it makes things slightly easier on the user
(no need to instantiate PBCs by hand) and disables BOX commands that a
user shouldn't use with periodic boxes. The old method still needs
ripped out and the counting code needs updated.