Convert PBC coordinates when dumping viz data.
The non-traditional (mirrored geometry) form of PBCs uses a relative
coordinate system, which the viz data made use of. The effect of this
was that it looked as if all molecules were diffusing within the
periodic boundaries. By using the relative XYZ coordinates, the periodic
box number, and the boundary locations, the "real" coordinates are
now written out to the viz data.

This feature could be made optional, but for now it is always applied.
Also, it might make more sense to add the periodic box number to the viz
data and let CellBlender handle it. This would have the added benefit of
allowing the user to toggle it on and off after the simulation was run.