Fix counting reactions in periodic boxes.
Counting reactions in periodic boxes (e.g. COUNT[vm,Scene.box,[1,0,0]])
was not working correctly. It would instead report the count in all the
boxes. The new behavior is still not 100% right (even though it's an
improvement), since it uses the periodic box of one of the reactants
instead of basing the PB off the location of the reaction.

Also, some additional check have been put in place to prevent illegal
types of counts. Specifically, it is illegal to use a count like
COUNT[vm,Scene.box] when PERIODIC_TRADITIONAL is set to FALSE. This
forces the user to specify a periodic box (e.g.[1,0,0]).