Add function to convert from PBC coordinate system
Added a function to convert from the relative PBC coordinate system to
the normal/absolute coordinate system. When using the non-traditional
form of PBCs, every molecule has not only an XYZ position but also a
periodic box that it inhabits. Using these two attributes along with the
position and dimensions of the box, we can convert the relative position
to an absolute one.

Currently, this function is only being used for viz data, but it will
likely be needed for TRIGGERs.