Bugfix: error placing surface molecule.
When placing surface molecules via place_surface_molecule, MCell tries
to find the closest place on a triangle to put the molecule by calling
closest_interior_point. If the molecule is on the line, it will move the
molecule toward the centroid. It determines if it is on the line by
checking if the cross product of the wall vertices and molecule position
(in uv coordinates) are distinguisable from zero. It was possible to
pass this test (e.g. a1 = -1e13 is indistinguishable), but then fail
later when trying to convert the uv coords to the grid index
(point_in_triangle_2D called from uv2grid) because the cross products
((pab and pbc) or (pab and pca)) technically had difference signs, which
indicates that the point is outside of the triangle. To avoid this
problem, we now simply call point_in_triangle_2D inside of
closest_interior_point. It would be best to only do this test once, but
this will likely require a much more extensive refactoring.