Add CMake file for DG branch.
The existing build system can cause problems for the DG branch because
of the DG parser. These problems could likely be fixed, but it's
probably easier to just switch to the cmake build system that is already
working in the master branch. This commit just adds the CMakeLists.txt
file to this branch (with the additional source files added in).