Bugfix: leaks caused by previous commit.
The previous commit would sometimes lead to molecules leaking out of the
mesh. This was particularly apparent in rapidly shrinking boxes or
cuboids, where large number of molecules ended up in the edges of the
box. We would move the molecule away from closest triangle along the
normal, but if the molecule was on the edge of a 90 degree angle, it
would just be pushed onto the neighbor's wall leaving it an ambiguous
position.

Note: this fix probably doesn't help in the case where you have an
object shrinking with really sharp edges.