Let user choose DG placement method.
Now, the user can choose whether volume molecules are placed via the old
(random location next to a wall) or new (closest position to the wall)
method. The keyword is DYNAMIC_GEOMETRY_RANDOM and it can be set to
either ON or OFF. This is kind of a stupid non-descriptive name, so we
should probably change it to something better in the future. The default
is ON.