Fix leaks when producing vm near close walls.
When placing a volume product created by a reaction involving a surface
molecule, the volume molecule would be bumped a small distance away from
the surface molecule's wall (along the normal). If another wall was
really close, which could easily happen if an sm produced a vm in a
sharp corner, then the vm would be bumped *across* the wall (i.e. it
would leak out). We get around this by ray tracing that small "bump"
distance, instead of just jumping to it.