Fix additional leak.
It's hard to be certain, but the removed code seemed unnecessary and
could cause leaks if a user was counting volume molecules in a region
with very sharp corners. The removed code was likely there to either
ensure that hits were counted correctly or possibly to prevent leaks
from occurring (which obviously didn't work). After removing this code,
all the tests in the nutmeg testsuite still run correctly. Granted, the
nutmeg testsuite is not thoroughly exhaustive, so there could be
problems with this fix. Additional tests (and benchmarks) should be
done.