Bugfix: backward rate not being used.
When using time-varying rate constants with a reversible reaction, the
reverse rate wouldn't be used. Instead it would just use the forward
rate for the reverse rate. Take the following reaction as an example:

vm + vm <-> vm [>"rxn1.txt", <"rxn2.txt"]

To break it down, the forward rate is this:

vm + vm -> vm ["rxn1.txt"]

And the backward rate should be this:

vm -> vm + vm ["rxn2.txt"]

But instead it would ignore "rxn2.txt" and just use "rxn1.txt" like
this:

vm -> vm + vm ["rxn1.txt"]

Note: this only seems to apply for the time-varying case, so you
wouldn't be affected with something like this:

vm + vm <-> vm [>1e4, <5e4]