High rxn probability is now an actual warning.
If any of the reaction probabilities were high, there would be a message
printed via mcell_log alerting the user along with the other final
statistics. This messaged looked like any other warning, but it wasn't
actually treated like the other warnings in MCell which use mcell_warn
and are directed to stderr.