Count SMs on regions in specific periodic boxes.
This commit makes it possible to count surface molecules in specific periodic
boxes in the same way we do with volume molecules already. For instance, if you
wanted to count all the "sm" molecules on the "sr" region of "Plane" in
periodic box [1,0,0], then you would simply use the following count statement:

{COUNT[sm,Scene.Plane[sr],[1,0,0]]}=> "./sm.100.dat"

The current implementation is a bit clunky and should be refactored.

Suport for counting on the ALL region or molecules enclosed in a region hasn't
been added yet.