Bugfix: miscounting surface molecules on regions.
Surface molecules were being miscounted for a couple of reasons. First,
we weren't reseting the counts or the molecule's periodic box location
in the case that an edge hit was reported as ambiguous from
find_edge_point. Secondly, the function to change a periodic box
(change_boxes_2d) was setting it incorrectly.