Refactor ray_trace_2d.
There are two long sections of code used to check for
absorption/reflection at region borders which are very similar to one
another, but just different enough that it's hard to combine them into a
single function. I've pulled out the first bit of code into a function
It doesn't help with the redundancy problem, but at least ray_trace_2d
isn't quite so long and convoluted now.