More refactoring of ray_trace_2d.
This "finishes" the second part of the previous refactor related to
checking for absorption/reflection at region boundaries. The previous
one was for inside-out movements. This one is for outside-in movements.
Both functions are pretty similar, so it would be nice to consolidate
them (or simplify the code in some way) if at all possible. Currently,
they both suffer from a serious case of parameter hell. I'm going to
address this in a future commit.