Bugfix: not handling SMs that can't be placed.
When you try to place a surface molecule at a specific location using
insert_surface_molecule, there's a possibility that the molecule can't
be placed (e.g if the grid is occupied). We weren't accounting for this
possibility when a molecule moved across a traditional PB, which in turn
could easily lead to a crash. Additionally, counting was also not being
handled correctly.