Better movement of SMs with trad PBCs.
This fixes some bugs and accuracy problems when placing and moving
surface molecules as they hit a (traditional) periodic boundary.
Previously, a new molecule was being created via
insert_surface_molecule. This had a number of problems associated with
it including efficiency, lack of proper cleanup, unnecessary molecule id
changes, etc. A separate problem was that displacements weren't being
updated correctly. Both of these seem to be working correctly in the
traditional case now. There may still be problems with counting and
displacement of SMs across non-traditional PBs still needs to be fixed.