Bugfix: over-counting molecules.
This is in response to a bug that Bob K reported. It seems to have been
caused by the recent refactoring of diffuse_3D. The problem, as far as I
can tell, was that they we were reseting the list of tentative
collisions, so we over-counted the number of molecules if they passed
through transparent walls. Note: this bug is likely also in the
periodic_BC_ng branch and needs to be fixed there.